[add] get board list method (wip)

+++ src/main/java/com/se/kumbangapiserver/dto/BoardDetailDTO.java
@@ -9,6 +9,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -53,7 +54,7 @@ public class BoardDetailDTO {
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime removedAt;
 
-    private Double distance;
+    private BigDecimal distance;
     private Details details;
 
     private List<FileDTO> files;
